structure updated
